LPS-9115 Make sure only people with UPDATE+Sharing permissions can manage
Previously, users would only see sharing entries created by them, so there was
no need to have a special "manage collaborators" permission.

With the behaviour change (now users see all collaborators, not only the ones
they shared with directly) we need to introduce the concept of collaborator
manager. Collaborators manager will be:

1. Aministrators,
2. The owner of the model (e.g. Document), or
3. Any user with UPDATE and "share" permission.

We cannot add a new "MANAGE" permission to SharingEntryAction, as this is not
a real permission, but a combination of factors external to the permission
system (e.g. the "shareable" check).